scripts/requirements.txt
# Python dependencies for Gemini Translation Script
# Install with: pip install -r requirements.txt

# Google Generative AI SDK for Gemini API
google-generativeai>=0.3.0

# Async PostgreSQL driver for database operations  
asyncpg>=0.29.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Enhanced logging and progress bars
rich>=13.0.0
tqdm>=4.66.0

# Optional: Data validation and serialization
pydantic>=2.5.0

# Optional: Enhanced async utilities
aiofiles>=23.2.1